pyproject.toml
[project]
name = "biterate-social-media-agent"
version = "0.1.0"
description = "AI agent for generating and posting BiteRate social media content"
requires-python = ">=3.9"
dependencies = [
    "notion-client>=2.2.1",
    "openai>=1.3.0",
    "mastodon.py>=1.8.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1",
    "pydantic>=2.0.0",
    "replicate>=1.0.0",
    "requests>=2.31.0",
    "python-telegram-bot>=22.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlite-vec>=0.1.0",
    "fastembed>=0.2.0",
    "numpy>=1.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
